fix: Xcode 26.4 / Clang 16 compatibility patches
- ASThread.h: suppress thread-safety-analysis on lock()/unlock() wrappers
- ASDisplayNode.mm: fix designated initializer field order
- FLAnimatedImage.m: cast enum to CGFloat for comparison
- TGVideoCameraPipeline.m: cast enum operands to uint32_t for bitwise OR
- TGVideoCameraGLRenderer.m: fix misleading indentation (tabs vs spaces)
- TGPhotoVideoEditor.m: add braces to fix misleading indentation warning
- GPUImageFramebuffer.m, PGPhotoCustomFilterPass.m, PGPhotoEditorPicture.m: same enum cast fix
- candidate.h (WebRTC): remove lifetimebound attr from void setter
- .bazelrc: add --disableExtensions for dev builds
- my_config.json: correct Team ID to MRPTPU87VT

+++ submodules/LegacyComponents/Sources/GPUImageFramebuffer.m
@@ -381,7 +381,7 @@ void dataProviderUnlockCallback (void *info, __unused const void *data, __unused
         
         if ([GPUImageContext supportsFastTextureUpload])
         {
-            cgImageFromBytes = CGImageCreate((int)_size.width, (int)_size.height, 8, 32, CVPixelBufferGetBytesPerRow(renderTarget), defaultRGBColorSpace, kCGBitmapByteOrder32Little | kCGImageAlphaPremultipliedFirst, dataProvider, NULL, NO, kCGRenderingIntentDefault);
+            cgImageFromBytes = CGImageCreate((int)_size.width, (int)_size.height, 8, 32, CVPixelBufferGetBytesPerRow(renderTarget), defaultRGBColorSpace, (uint32_t)kCGBitmapByteOrder32Little | (uint32_t)kCGImageAlphaPremultipliedFirst, dataProvider, NULL, NO, kCGRenderingIntentDefault);
         }
         else
         {
